WIP factor out floating toolbar logic from selection toolbar and link menu
